empty project

+++ app/controllers/simulator_controller.rb
@@ -60,12 +60,23 @@ class SimulatorController < ApplicationController
     @project.author = current_user
 
     data_url = params[:image]
-    jpeg      = Base64.decode64(data_url['data:image/jpeg;base64,'.length .. -1])
-    image_file = File.new("preview_#{Time.now()}.jpeg", "wb")
-    image_file.write(jpeg)
+
+    str =data_url['data:image/jpeg;base64,'.length .. -1];
+    if str.to_s.empty?
+      image_file = File.open(Rails.root.join("app/assets/images/empty_project/no_preview.png"), "rb")
+
+    else
+      jpeg       = Base64.decode64(str)
+      image_file = File.new("preview_#{Time.now()}.jpeg", "wb")
+      image_file.write(jpeg)
+    end
+
     @project.image_preview = image_file
     @project.save
-    File.delete(image_file)
+
+    if !str.to_s.empty?
+      File.delete(image_file)
+    end
 
     # render plain: simulator_path(@project)
     # render plain: user_project_url(current_user,@project)
